Updated jenkinsfile.adoc, incorrect referral fixed
Hi!

The existing wiki page uses a non-existent reference in the "Parallel Execution" section, which refers to a "node" instead of an "agent" when describing the usage of multiple agents in one pipeline. I'm guessing some changes were made to this document, and the wording was changed from node to an agent which is why this reference is not working at the moment.