Recommend Ukrainian Red Cross for relief donations (#6622)
Recommend donations to Ukrainian Red Cross

https://github.com/jenkins-infra/jenkins.io/issues/6584 recommends that
we change from the International Committee of the Red Cross to the
Ukrainian Red Cross.

Governance board meeting 21 Aug 2023 agreed with the proposal and
assigned Mark Waite the action item to submit a pull request making the
change.

Board members will be mentioned in the pull request so that they can
comment on the pull request.