Include 2.414.1 in SECURITY-3188 list of releases with fix (#6631)
* Include 2.414.1 in SECURITY-3188 list of releases with fix

Can already be inferred by the reader since the preceding LTS, 2.401.3
includes the fix.  Adding the text makes it clear that the issue does
not affect LTS 2.414.1.

May want to consider removing the security advisory item from the
2.414.1 changelog.  It was a backport, but the backport did not result
in a user-visible change compared to 2.401.3.

* Add date to backported security advisory

* Fix case of the text

Co-authored-by: Daniel Beck <1831569+daniel-beck@users.noreply.github.com>

* Include Oxford comma

My preference to use the Oxford comma anyway.  Thanks for noting it

* Include 2.414.1 in LTS fix list for advisory 2023-07-26

Fix was backported to that release.  Make it consistent with the
2023-03-08 advisory

* Use same plural verb "encode" as is used earlier

---------

Co-authored-by: Daniel Beck <1831569+daniel-beck@users.noreply.github.com>
Co-authored-by: Kevin Martens <99040580+kmartens27@users.noreply.github.com>