Update links to Google Cloud and Oracle Cloud on home page (#6706)
* Point to a Google Cloud page that is not archived

Links should go to documentation that is actively maintained by the
cloud provider, not to an archive page that warns it may disrupt the
Kubernetes environment.

* Use more recent Oracle Cloud destination link

The previous link used the deprecated term "master".  That page has been
replaced on Oracle documentation, but the replacement is more specific
than this architecture reference. Since the title of the link includes
architecture, I updated it to point to the architecture page.