Remove link to private topic in community (#6748)
Remove link to private community thread

The organizers of the 2021 Hacktoberfest project to localize in French
are not participating in HAcktoberfest 2023 in Jenkins.  The thread in
community.jenkins.io was made private at the end of the 2021 effort.

The remaining links are useful and the French localization effort is
still a good effort to continue.