[JENKINS-72546] Include Amazon Linux 2 in RHEL 7 derivatives (#7017)
Include Amazon Linux 2 in RHEL 7 derivatives

https://issues.jenkins.io/browse/JENKINS-72546 is a recent request that
shows OpenSSH versions before 7.6 are still being used.  Assure that
readers know that Amazon Linux 2 is no longer supported because it is
a derivative of Red Hat Enterprine Linux 7.

Usually we don't revise blog posts, but this blog post is frequently
referenced.  I think it is better to revise the blog post than to post
repeated additions to the phrasing of the blog post.