Update the download verification instructions
The Signature algorithm used for the Jenkins war file has been
strengthened with the most recent update of the Jenkins war signing
ceritifcate.  Use the current string in the sample output.

The plugins site no longer has a tab called "Archives".  It is now the
"Releases" tab and does not include the checksums for the plugin releases.
Link to the update center page that includes the checksums for plugins.