Remove date dependent phrase from governance agenda link (#7246)
* Remove date dependent phrase from governance agenda link

Readers that need the archived copies of the agenda should not look
for them in HackMD.  We use HackMD as a real-time editing facility for
markdown of the upcoming meeting.  During the meeting, we use it to take
notes for the current meeting.  After the meeting, the notes are copied
to the archive locations.

* incoming -> upcoming

The governance meeting is approaching as a date, not as an object in flight