The hudson.model.DirectoryBrowserSupport.allowAbsolutePath is obsolete
hudson.model.DirectoryBrowserSupport.allowAbsolutePath will
be obsolete in the first weekly release after the merge of
https://github.com/jenkinsci/jenkins/pull/9387

Was not sure if this is the correct pattern to record a property that
is no longer available.  I tried to follow the pattern used by others
in the file.  Happy to adjust to whatever format or style is preferred.