Update container images versions in hello-world tutorial
It is not clear to me why the updatecli process did not propose changes
for these containers.  It is especially curious that some cases remained
very far out of date on older Alpine versions and at least one case had
an update in the declarative Pipeline example that was not matched in
the scripted Pipeline example.

Switches the Maven example to use Java 21.