Phrase 2.466 changelog consistently with 2.465
The 2.465 changelog correctly states that a breaking change in the
Apache Maven Release plugin caused the build to be "burned" (unavailable,
discarded).

The 2.466 changelog originally used less clear phrasing to describe
the rollback of the Apache Maven Release plugin as a workaround for the
breaking change.  This change updates the phrasing to clearly state that
the rollback of the plugin release is used to avoid the breaking change
in Apache Maven Release plugin 3.1.0.