Link to the OpenJDK upgrade issue in Jira
https://issues.jenkins.io/browse/JENKINS-72665 has more details for
readers that need to better understand the root of the issue and the
reason that they must restart the Java process when the Java runtime
is upgraded.

Segmentation violation or segfault is reported by Java when Java tries
to spawn a new process after an upgrade from a Java 11 version before
11.0.22 to a Java 11 version 11.0.22 or later.  Same failure when they
upgrade Java from a Java 17 version before 17.0.10 to a Java version
17.0.10 or later.

The segfault is only reported when the previous version of the JDK is
trying to use the spawnhelper of the new version.  If the user stops
the Java process and restarts it, the problem will not occur.