Replace incorrect GCP link on downloads page, update GCP tutorial (#7610)
* Replace incorrect GCP link on downloads page, update GCP page

The link to GCP from the downloads page is for an unrelated page.
The original page was a 2018 blog post.  Much newer information is
available from other sources.  Rather than link to an unrelated page,
let's link to a page that we maintain and that we can update with more
content.

The linked blog post from medium.com that is now included in the Jenkins
tutorial page is the most thorough Jenkins installation tutorial (that
I could fine) for Google Cloud Platform.  It describes the Jenkins
controller installation and the configuration of Kubernetes agents for
that Jenkins controller using Google Kubernetes Engine.

The video "How to integrate Jenkins with GCP" provides detailed
instructions to configure and use the `gcloud` command from inside a
Jenkins job.  It does not require any knowledge of Packer or any use of
Packer images.

The second video, "Using Google Cloud Run", builds on the material from
the first video and shows how to use the serverless Google Cloud Run
facility to deploy a container image to Google Cloud.

The third video describes how to use Google Secret Manager for Jenkins
credentials.

* Suppress typos for video URL