Do not depend on nodeJS version in example output (#7695)
A recent documentation feedback noted that the incorrect version number
string is included in the sample output.  Rather than adjust the example
output, let's use example output that is unlikely to change between NodeJS
releases.  This example uses `process.platform` and `process.env.CI`.
Those are not likely to change from one version of nodeJS to the next.

https://docs.google.com/spreadsheets/d/1nA8xVOkyKmZ8oTYSLdwjborT0w-BpBNNZT0nxR9deZ8/edit?usp=sharing
is the feedback sheet that included the note that the NodeJS version
number is incorrect in the example output.