Use broader description of RHEL and its derivatives
Use the precise name of Red Hat Enterprise Linux and refer to the
derivatives in the title rather than listing the derivatives.  Include the
names of key derivative operating systems in the body text so that search
engines continue to find them.

The Red Hat usage guidelines are very clear that they operating system is
named Red Hat Enterprise Linux and should not be referenced as "Red Hat".