Link from download pages to install guides, not pkg.jenkins.io
The pkg.jenkins.io site is the download site, but it is not the best
site for installation instructiosn for Linux packages.  Users of the
Linux packages do not use a download URL.  They use a series of steps
that are specific to their operating system and its package manager.
The link to the download site is a link to installation steps listed
on pkg.jenkins.io.  The installation steps on www.jenkins.io are more
thorough and include specific instructions for Fedora Linux that are
not available on pkg.jenkins.io

Adds Fedora Linux as a seperate link in the list of packages because it
intentionally has separate instructions for its `dnf` package manager.

Red Hat Enterprise Linux 10 is planned to include the `dnf` package
manager when it releases in mid-2025.  When it releases, we'll need to
revisit the instructions.