Explain the JCasC behavior on deprecated configuration elements
Configuration as code intentionally cancels Jenkins startup when a
deprecated configuration element is detected.  This forces administrators
to resolve the deprecations immediateely rather than delyaing the
resolution of the deprecation.

Describe the configuration as code setting that will change the
default behavior to warn on deprecation rather than canceling startup.
Explain the risks of that choice.

https://github.com/jenkinsci/jenkins/pull/9903#issuecomment-2640241966
is the pull request comment that requests a more detailed description
of the behavior.