Move GSoC to the top, archive Blue Ocean, mark k8s done
Google Summer of Code is a very active project in Jenkins.  Let's
place it at the top of the list of projects.

Blue Ocean is only receiving selective updates for significant
security issues or functional defects.

The Kubernetes documentation project was completed in 2020 Google
Season of Docs.  Kubernetes documentation updates are now part of the
regular efforts of the Jenkins documentation special interest group.