Create LTS 2.504.1 changelog and upgrade guide (#8033)
* create 2.504.1 changelog and upgrade guide

* add reference PRs for changelog entry

* update changelog entry status

* expand NPE for full words

* Remove the developer entries from 2.504.1 changelog

While the developers considered it a major RFE to add a new ExtensionPoint
to receive events from CLI executions, end users will not see any
significant changes in this release due to that change.

Allowing all immutable List subclasses from Java 11 is likely undetectable
for almost all users.

* Provide more explanation text for OKLCH color change

* More explanation of the task notification fix

* Simplify deprecated plugin sorting message

* Winstone 8.5 was included in Jenkins 2.492.3

The Winstone 8.5 release already included the fix for user names with
backslash.

* Remove the www.jenkins.io prefix for easier development

When we include the full URL, it makes the development environment
switch to the production site when the link is clicked.

* apply suggestions from @timja

* fix link code

* Update content/_data/changelogs/lts.yml

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>

---------

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>