Fix the example Docker plugin installation definition
The Docker install from https://www.jenkins.io/doc/book/installing/docker/
fails to load plugins.  The container starts but does not load many of
the plugins because the token macro plugin fails to load.  The token
macro plugin fails to load because the plugin installation manager tool
does not download its dependency on the json path api plugin.

https://github.com/jenkinsci/plugin-installation-manager-tool/issues/628
reports the issue in the plugin installation manager tool.

The fix was already applied to the Blue Ocean tutorial
https://www.jenkins.io/doc/tutorials/create-a-pipeline-in-blue-ocean/ in:

* https://github.com/jenkins-infra/jenkins.io/pull/7084

Testing done

Checked that the instructions work as described and created a Jenkins
job that ran docker commands from inside the controller.