Remove low impact bug fix from 2.508 changelog (#8081)
Address concern from Daniel Beck that the spelling fix in the message
is not impactful enough to justify an entry in the weekly changelog.

The Jenkins 2.478 changelog has an empty entry that is displayed as
"No notable changes in this release".  This will be displayed in the
same way.

For comparison, see 2.478 on www.jenkins.io and on the GitHub repository:

* https://www.jenkins.io/changelog/2.478/
* https://github.com/jenkinsci/jenkins/releases/tag/jenkins-2.478

This changelog retains the comments for the skipped entries.  The
2.478 changelog data has not been adjusted to include the skipped
entries.