Add scarf.sh tracking pixel as test for several authors (#8252)
As discussed in Docs Office Hours 10 July 2025, the Linux Foundation
has an agreement with scarf.sh for page tracking and download tracking.
We'd like to test drive this tracking through our fast.ly content delivery
network to see if the data that is collected will help us understand
page use.